Implement some of the boilerplate of a make-shift game library.
While this is not based on U3, it should be good starting point for starting to port some of the older Crayon games to PlexiOS using a CommonScript runtime. Later updates can be migrated fairly easily and will help just get things off of old Crayon or JavaScript or intermediate formats.